discovering errors in BSA extraction, noting with comments
